Moved Prefabs (last time), Changed Super Buttons, Added Pause Menu Alert
Moved the prefabs around for the last time. Changed the Ricky Bobby and Keanu Reeves buttons so that they take the player right back to the game after being pressed. Added an alert when the Ricky and Keanu buttons are active, directing player to the pause menu. This is the last update for this game (unless I change my mind).